# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "CustomFieldValueOut": {
        "properties": {
          "field_definition_id": {
            "title": "Field Definition Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "field_definition_id",
          "label",
          "value"
        ],
        "title": "CustomFieldValueOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoleCreateRequest": {
        "properties": {
          "can_create_settings": {
            "default": false,
            "title": "Can Create Settings",
            "type": "boolean"
          },
          "can_delete_settings": {
            "default": false,
            "title": "Can Delete Settings",
            "type": "boolean"
          },
          "can_edit_settings": {
            "default": false,
            "title": "Can Edit Settings",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RoleCreateRequest",
        "type": "object"
      },
      "RoleOut": {
        "properties": {
          "can_create_settings": {
            "title": "Can Create Settings",
            "type": "boolean"
          },
          "can_delete_settings": {
            "title": "Can Delete Settings",
            "type": "boolean"
          },
          "can_edit_settings": {
            "title": "Can Edit Settings",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "can_create_settings",
          "can_edit_settings",
          "can_delete_settings",
          "is_archived"
        ],
        "title": "RoleOut",
        "type": "object"
      },
      "RoleUpdateRequest": {
        "properties": {
          "can_create_settings": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can Create Settings"
          },
          "can_delete_settings": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can Delete Settings"
          },
          "can_edit_settings": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Can Edit Settings"
          },
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "RoleUpdateRequest",
        "type": "object"
      },
...
      "TagCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TagCreateRequest",
        "type": "object"
      },
      "TagOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TagOut",
        "type": "object"
      },
      "TagUpdateRequest": {
        "properties": {
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TagUpdateRequest",
        "type": "object"
      },
      "TagsUpdateRequest": {
        "properties": {
          "tag_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Tag Ids",
            "type": "array"
          }
        },
        "required": [
          "tag_ids"
        ],
        "title": "TagsUpdateRequest",
        "type": "object"
      },
...
      "TeamOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TeamOut",
        "type": "object"
      },
...
      "TicketOut": {
        "properties": {
          "age_seconds": {
            "title": "Age Seconds",
            "type": "integer"
          },
          "business_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Id"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryOut"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserOut"
          },
          "custom_field_values": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CustomFieldValueOut"
            },
            "title": "Custom Field Values",
            "type": "array"
          },
          "customer": {
            "title": "Customer",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "doctor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Name"
          },
          "first_response_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Response At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mobile_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile Number"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "sla_breached": {
            "title": "Sla Breached",
            "type": "boolean"
          },
          "sla_due_at": {
            "format": "date-time",
            "title": "Sla Due At",
            "type": "string"
          },
          "sla_hours": {
            "title": "Sla Hours",
            "type": "integer"
          },
          "sla_remaining_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Remaining Seconds"
          },
          "slack_channel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Channel Id"
          },
          "slack_message_ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Message Ts"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "support_assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TagOut"
            },
            "title": "Tags",
            "type": "array"
          },
          "team": {
            "$ref": "#/components/schemas/TeamOut"
          },
          "ticket_number": {
            "title": "Ticket Number",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticket_number",
          "title",
          "description",
          "customer",
          "business_id",
          "mobile_number",
          "doctor_name",
          "category",
          "team",
          "priority",
          "status",
          "sla_hours",
          "sla_due_at",
          "owner",
          "support_assignee",
          "created_by",
          "slack_channel_id",
          "slack_message_ts",
          "created_at",
          "first_response_at",
          "resolved_at",
          "closed_at",
          "updated_at",
          "sla_breached",
          "sla_remaining_seconds",
          "age_seconds"
        ],
        "title": "TicketOut",
        "type": "object"
      },
      "TicketPriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "TicketPriority",
        "type": "string"
      },
      "TicketStatus": {
        "enum": [
          "open",
          "in_progress",
          "pending",
          "resolved",
          "closed"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
...
      "UserOut": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slack_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack User Id"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Support Operations Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/roles": {
      "get": {
        "operationId": "list_roles_api_roles_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleOut"
                  },
                  "title": "Response List Roles Api Roles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Roles",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "create_role_api_roles_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/roles/{role_id}": {
      "patch": {
        "operationId": "update_role_api_roles__role_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Role",
        "tags": [
          "roles"
        ]
      }
    },
...
    "/api/tags": {
      "get": {
        "operationId": "list_tags_api_tags_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagOut"
                  },
                  "title": "Response List Tags Api Tags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tags",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "create_tag_api_tags_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}": {
      "patch": {
        "operationId": "update_tag_api_tags__tag_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "title": "Tag Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Tag",
        "tags": [
          "tags"
        ]
      }
    },
...
    "/api/tickets/{ticket_id}/tags": {
      "post": {
        "operationId": "update_ticket_tags_api_tickets__ticket_id__tags_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ticket Tags",
        "tags": [
          "tickets"
        ]
      }
    },
...
  }
}